Artifact manager refactoring
