+++ src/artifacts_repository.rs
@@ -14,8 +14,5 @@
    limitations under the License.
 */
 
-pub mod handlers;
-pub mod model;
-
-pub use crate::artifacts_repository::artifact_manager::*;
-pub use crate::artifacts_repository::hash_util::*;
+pub mod artifact_manager;
+pub mod hash_util;
