+++ src/docker/v2/handlers.rs
@@ -16,6 +16,6 @@
 
 pub mod blobs;
 pub mod manifests;
-pub use crate::artifact_manager::HashAlgorithm;
+pub use crate::artifacts_repository::hash_util::HashAlgorithm;
 
 pub use crate::node_manager::*;
