+++ src/lib.rs
@@ -18,7 +18,7 @@
 
 extern crate lazy_static; // Must be done in crate root
 
-pub mod artifact_manager;
+pub mod artifacts_repository;
 pub mod docker;
 pub mod document_store;
 pub mod metadata_manager;
